[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.04 ms, Memory: 73.1 MB -BaekjoonHub

+++ 프로그래머스/lv1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.java
@@ -3,7 +3,7 @@ class Solution {
         long[] answer = new long[n];
         
         for(int i=1;i<=n;i++) {
-            answer[i-1] = (long)x*(long)i; 
+            answer[i-1] = (long)x*i; 
         }
         
         return answer;
